Properly handle republishing, plus a small refactor
